requirements.txt
faiss-cpu
google-genai
pypdf
python-docx
markdown
beautifulsoup4
numpy
nltk
python-dotenv
